[squash] review comments. Moving the session_cloud stuff into a fixture like a sane person would do
